Update current figma icons file name version (#412)
* Update name of current Figma icons library

Updated the name of the current Figma icons library to include Astro 6 & 7 in the title in preparation for a new major vesion icon library. This change bumps the version of the Figma file to 1.1.0

* Update designers.md

Updated the icons library file name on the for designers page

* Update developers.md

Updated name of icons file in the For Developers page.

* Update downloads.astro

Updated the icon library name to Astro 6 & 7 in the Downloads page.